move formatting value into yaxisformat module too
